Change f64 to f32 to reduce memory usage

+++ src/any/delayed.rs
@@ -8,10 +8,10 @@
 /// [Delayed](./struct.Delayed.html) object.
 ///
 /// An argument object can be constructed either from [a single value](./struct.Arg.html#impl-From<%26%27a%20T>) or
-/// [a value and a timeout](./struct.Arg.html#impl-From<(%26%27a%20T%2C%20f64)>).
+/// [a value and a timeout](./struct.Arg.html#impl-From<(%26%27a%20T%2C%20f32)>).
 #[derive(Copy, Clone, Debug)]
 pub struct Arg<'a, T: ?Sized> {
-  timeout: f64,
+  timeout: f32,
   value: &'a T,
 }
 
@@ -24,14 +24,14 @@ impl<'a, T: ?Sized> From<&'a T> for Arg<'a, T> {
   }
 }
 
-impl<'a, T: ?Sized> From<(&'a T, f64)> for Arg<'a, T> {
+impl<'a, T: ?Sized> From<(&'a T, f32)> for Arg<'a, T> {
   /// Performs the conversion from a `value` and a certain amount of `timeout` to be elapsed before the new `value` is
   /// being assigned and held.
   ///
   /// # Panics
   ///
   /// Panics if `timeout` is not a number or less than or equal to 0.0 .
-  fn from((value, timeout): (&'a T, f64)) -> Self {
+  fn from((value, timeout): (&'a T, f32)) -> Self {
     debug_assert!(
       timeout > 0.0,
       "timeout must be a number that is greater than 0.0!"
@@ -45,8 +45,8 @@ impl<'a, T: ?Sized> From<(&'a T, f64)> for Arg<'a, T> {
 /// A `Delayed` object can be constructed by calling [Delayed::new](./struct.Delayed.html#method.new) function.
 #[derive(Copy, Clone, Debug)]
 pub struct Delayed<'a, T: ?Sized> {
-  t: f64,
-  timeout: f64,
+  t: f32,
+  timeout: f32,
   pending_value: &'a T,
   value: &'a T,
 }
@@ -156,7 +156,7 @@ impl<'a, T: ?Sized> Delayed<'a, T> {
   /// // The latest value has been assigned since the `delayed_value` has timeout.
   /// assert_eq!(delayed_value.get_value(), &1);
   /// ```
-  pub fn step(&mut self, dt: f64) {
+  pub fn step(&mut self, dt: f32) {
     debug_assert!(dt >= 0.0, "dt must be a positive number!");
     self.t += dt;
     if self.t < self.timeout {
